requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
PyMuPDF==1.18.19
google-cloud-vision==2.7.1
openai==0.10.2
gunicorn==20.1.0
Werkzeug==2.0.2